Change regex state register to vector from hash map.
Vector is faster and simpler since we anyway can reference the states
by a integer. This means we can have the id of the states be simply
their index in the vector, so that state get is still O(1).
While removal will now be slower, this is not really a problem, as the
application will not have a need to remove really any states from the
register.