Regex: Avoid allocation of String in NFA bracket states
Instead make them take a `&str` reference into the regular expression
string. This is a good performance improvement as it avoids
unnecessary allocation. The cost is only that we now need to make sure
that the regex string has a longer lifetime than the NFA states.
Since the regex anyway has to be fetched earlier than the states are
initialized this should not be a problem to ensure.