Refactor: Use `if let` over `match` when only caring about one value
Nicer to use `if let` if we only care about one pattern in the
enum. Also change the `insert_or_follow_split` function so that it
will not attempt to unwrap `None` values in the out arrows for
states. Such values should not be there when we reach that function,
but that should be ensured by another, earlier check.