Return index of regex match end, and implement greedy regex matching
In the NFA simulate function, which matches a input string to the
regex that has built the NFA. For later functionality we need to know
where the matching substring of the input string ends. This will then
depend on whether we match with the greedy strategy or not, so we
implement an option for this in the NFA simulate function.