Go directly from infix regex to NFA.
Before we first converted the infix regex to a regex in the postfix
notation, then converted that to a NFA. These two steps can easily be
combined into just one iteration over the regex characters. This is
faster, due to having to iterate only once. More importantly it
enables the code to be shorter, which makes it easier given that the
logic itself for both steps somewhat involved.

The old functions are not removed quite yet, but all tests are now
migrated to the new function.