Regex matching for substrings, main entry-point for module
A new `Regex` struct that will be the main entrypoint to the
regex module. Matching that can start at any position inside the
string, as opposed to only from the start. Also implement now the
caret (^) and dollar ($) characters for matching the start and end of
the input string (assumed to be a single line).