Error handling of invalid operator usage in regex
Handle errors caused by regex operators not having enough operands to
act on. This can be caused by forgetting to escape a operator or
exchanging the order of two symbols in the regex. The exact cause is
hard to pinpoint, but in any case it is caused by a regex that is
malformed by the user in some way.
The exception to this is the concatenation operator which is implicit
in the user regex and only used as a explicit operator
internally. Therefore it is a bug if it is presented with too few
operands, not a user error.