Be more willing to match regex ending in `$` (until end)
In this case we are greedy on the tail of the match even if greedy
matching is not specified. This makes the pattern more likely to match
a given line, so we are effectively more willing to match a regex that
says we should go to the end of the line.