Match greedily by default.
This is the way it is done by e.g. `grep`. Also this is much nicer for
us because matching the start of the regex non-greedily would add a
lot of logic and have a performance penalty. Instead we have a `lazy`
argument that enables lazy matching for the tail of the regex, giving
a performance boost since the simulation loop can end early.