Fix bug related to precedence of groupings `(`, `)`.
This bug showed itself as a panic when a grouping contained a
alteration and at least a concatenation or other operator. Since the
alteration operator has the lowest precedence, everything was popped
from the operator stack until the opening parenthesis was
reached. That was panic'ed in the precedence check due to it not being
a valid regex operator (which is because groupings are handled in
their own way). The solution is that in our implementation a opening
paren in the operator stack should never be popped, unless by the
closing paren. Therefore added a check for the opening paren in the
function determining when to pop the operator stack.