Restructure and cleaning of NFA simulation function
A major clean up of the NFA simulation function, which was rather
large and with hard to parse logic. The approach chosen is to use
two new objects representing a simulation of a NFA and the next states
for any given step in the simulation.
Also reduces the amount of comment, which are now less needed due to
more named functions and objects.
Outward facing functionality should be unchanged, though the
simulation is a tiny bit slower than before.