Optimization for regex beginning with literal string: find it first
First searching for the literal string in the input is a great
improvement in performance in the quite common case that the regex
begins with a literal string.

A side-effect of this change is that support for lazy matching on the
tail is dropped. This due to the amount of extra logic needed to still
support this with the current implementation.