Optimization for regex with a literal string tail
First reverse-search the input for the string in question, and use
the result of that search to narrow the input needed to check with the
NFA. Implementing this required some restructuring of the substring
matching functionality of the regex module. The result is much faster
matching in this special case. The result will be even more pronounced
if we also implement the "not going line by line" optimization later.