Fix bug where str slice does not hit code-point boundary.
This happened in relation to the literal tail optimization. In that we
after NFA matching check if the substring after the match with the
length of the tail is equal to the tail. However when that length in
bytes placed us not at a code-point boundary the code paniced. It
doesn't need to: if we get into that state there is no match.