Make `count` work in recursive searches, change operation function
Making counting matching lines work in recursive searches. In that
process: make the match select the function, then pass that as an
argument, instead of having the match being run on every line. This
should be somewhat faster, since the match is run more rarely. Also it
is more concise.