Fix handling of matching on the empty string.
Fix some inconsistencies in how matches on the empty string was
handled.

Consider the regex `a*`. It clearly matches the string `b` because it
matches the empty string at the beginning, containing zero a's. This was
not the case in the implementation before this commit.

Turns out this simplified the NFA simulate function as well.