Fix and simplify handling of literal head/tail optimization
This commit fixes several bugs related to the implementation of this
optimization, most of which are tested now in the added unit-tests.

The strategy has been to go back to a simpler, more readable
implementation that is easier to reason that is correct. This actually
comes at only a small cost to performance.

Lesson learned: Focus on great unit-test coverage before attempting to
implement any sort of optimization!